+++ Microcontrollers/Eindopdracht/ultrasonic_sensor.h
@@ -18,7 +18,7 @@ void ultrasonic_send_pulse();
 /* get the value of the timer which equals time taken for sound to		*/
 /* return    			                                                */
 /************************************************************************/
-uint16_t ultrasonic_get_timer_dist();
+float ultrasonic_get_timer_dist();
 /************************************************************************/
 /* handle for the interrupt                                             */
 /************************************************************************/
